requirements.txt
aniso8601
astor
attrs
Babel
bce-python-sdk
certifi
charset-normalizer
click
colorama
contourpy
cycler
decorator
dill
Flask
Flask-Babel
Flask-Cors
flask-restx
fonttools
future
idna
imageio
imgaug
importlib-metadata
itsdangerous
jamo
Jinja2
joblib
jsonschema
kiwisolver
lmdb
MarkupSafe
matplotlib
multiprocess
networkx
nltk
numpy
opencv-python
opt-einsum
packaging
pandas
Pillow
pip
protobuf
pyclipper
pycryptodome
pyparsing
pyrsistent
python-dateutil
pytz
PyWavelets
PyYAML
regex
requests
scikit-image
scipy
setuptools
Shapely
six
tifffile
tqdm
urllib3
visualdl
Werkzeug
wheel
wincertstore
zipp
